# ENH: port np.core.overrides to C for speed