… benchmarks/benchmarks/bench_overrides.py
… doc/neps/nep-0018-array-function-protocol.rst
… numpy/core/src/multiarray/arrayfunction_override.h
… numpy/core/src/multiarray/methods.c
… numpy/core/src/multiarray/multiarraymodule.c
… numpy/core/src/multiarray/multiarraymodule.h
… numpy/core/tests/test_overrides.py
… numpy/core/tests/test_shape_base.py
… tools/travis-test.sh