**TL,DR:** This significantly speeds up dispatch for `__array_function__`, reducing the overhead down to about 0.7 microseconds for typical cases (vs 2-3 microseconds currently with a pure Python implementation). For functions that handle many arguments (like `concatenate`), this reduces the overhead of dispatching by something like 100x, i.e., from "makes concatenate 4x slower" to "barely unnoticeable".

Original post:

-----------------------------------------------

Still needs:
- [x] documentation for the C functions
- [x] actually using the C functions in np.core.overrides
- [x] some way to test get_overloaded_types_and_args directly (a Python wrapper for testing?)

Currently NumPy doesn't even import properly after I build it with this change. Hopefully I'm doing something obviously wrong!
```
Original error was: dlopen(/Users/shoyer/dev/numpy/build/testenv/lib/python3.7/site-packages/numpy/core/_multiarray_umath.cpython-37m-darwin.so, 2): Symbol not found: _array_function_implementation_or_override
  Referenced from: /Users/shoyer/dev/numpy/build/testenv/lib/python3.7/site-packages/numpy/core/_multiarray_umath.cpython-37m-darwin.so
  Expected in: flat namespace
 in /Users/shoyer/dev/numpy/build/testenv/lib/python3.7/site-packages/numpy/core/_multiarray_umath.cpython-37m-darwin.so
```

Also: I expect it's obvious, but I'm pretty new to C and Python's C-API. I expect there are lots of ways this PR could be improved.