@shoyer 2018-11-03
I've made some progress -- NumPy builds and imports, but running the tests in numpy/core/tests/test_overrides.py segfaults.

I'm currently trying to get gdb and python configured so I can get a stacktrace...