@charris 2018-11-03
This seems a bit premature. Is there an actual performance benefit from doing this?