@shoyer 2018-11-03
> This seems a bit premature. Is there an actual performance benefit from doing this?

Well, it doesn't work yet so I haven't been able to benchmark how much it helps yet. But checking for overrides definitely adds some overhead, and it's now on every single NumPy function call, e.g.,
```
In [3]: x = np.arange(10)

In [4]: %timeit np.sum(x)
5.53 µs ± 56.9 ns per loop (mean ± std. dev. of 7 runs, 100000 loops each)

In [5]: %timeit x.sum()
1.83 µs ± 11.7 ns per loop (mean ± std. dev. of 7 runs, 1000000 loops each)

In [6]: %timeit np.sum.__wrapped__(x)
2.68 µs ± 40.4 ns per loop (mean ± std. dev. of 7 runs, 100000 loops each)

In [9]: %prun [np.sum(x) for _ in range(100000)]
         1500004 function calls in 0.907 seconds

   Ordered by: internal time

   ncalls  tottime  percall  cumtime  percall filename:lineno(function)
   100000    0.197    0.000    0.339    0.000 overrides.py:15(get_overloaded_types_and_args)
   100000    0.175    0.000    0.175    0.000 {method 'reduce' of 'numpy.ufunc' objects}
   100000    0.095    0.000    0.394    0.000 fromnumeric.py:1968(sum)
   100000    0.094    0.000    0.282    0.000 fromnumeric.py:69(_wrapreduction)
   200000    0.073    0.000    0.073    0.000 {built-in method builtins.hasattr}
   100000    0.057    0.000    0.858    0.000 overrides.py:148(public_api)
   100000    0.056    0.000    0.789    0.000 overrides.py:64(array_function_implementation_or_override)
        1    0.048    0.048    0.906    0.906 <string>:1(<listcomp>)
   100000    0.026    0.000    0.026    0.000 overrides.py:57(<listcomp>)
   100000    0.021    0.000    0.021    0.000 {method 'insert' of 'list' objects}
   100000    0.017    0.000    0.017    0.000 {built-in method builtins.isinstance}
   100000    0.013    0.000    0.013    0.000 {method 'append' of 'list' objects}
   100000    0.012    0.000    0.012    0.000 {method 'items' of 'dict' objects}
   100000    0.011    0.000    0.011    0.000 fromnumeric.py:1963(_sum_dispatcher)
   100000    0.010    0.000    0.010    0.000 {built-in method builtins.len}
        1    0.001    0.001    0.907    0.907 <string>:1(<module>)
        1    0.000    0.000    0.907    0.907 {built-in method builtins.exec}
        1    0.000    0.000    0.000    0.000 {method 'disable' of '_lsprof.Profiler' objects}
```

I'll check the ASV suite, but I suspect it would be a meaningful performance win if we shave off 1-2 us of overhead on every NumPy operation. Typical NumPy operations take ~10 us on small arrays.