@pv 2018-11-03
asv shows regressions on array_ufunc in a number of placs https://pv.github.io/numpy-bench/#regressions