@shoyer 2018-11-03
Here's a spreadsheet with all the changes that I identified as due to `__array_function__`:
https://docs.google.com/spreadsheets/d/15-AFI_cmZqfkU6mo2p1znsQF2E52PEXpF68QqYqEar4/edit#gid=0