@shoyer 2018-11-05
OK, I managed to debug this. For future reference, I wasn't able to get `gdb` working on OS X, but `lldb` worked pretty well.

It still needs docs and a bit of cleanup but it's in good enough shape to run benchmarks and for anyone interested to take a look at the implementation. Here are some highlights.

The general override benchmarks are only about 10% faster for functions that handle a single argument, but 3x faster for functions given a long list of arrays: 
```
## on master
[ 58.33%] ··· bench_overrides.ArrayFunction.time_mock_broadcast_to_duck                                    3.47±0.07μs
[ 66.67%] ··· bench_overrides.ArrayFunction.time_mock_broadcast_to_numpy                                   2.46±0.03μs
[ 75.00%] ··· bench_overrides.ArrayFunction.time_mock_concatenate_duck                                     4.24±0.08μs
[ 83.33%] ··· bench_overrides.ArrayFunction.time_mock_concatenate_many                                         271±9μs
[ 91.67%] ··· bench_overrides.ArrayFunction.time_mock_concatenate_mixed                                     6.52±0.1μs
[100.00%] ··· bench_overrides.ArrayFunction.time_mock_concatenate_numpy                                    3.09±0.01μs
## with this PR
[ 58.33%] ··· bench_overrides.ArrayFunction.time_mock_broadcast_to_duck                                    2.78±0.02μs
[ 66.67%] ··· bench_overrides.ArrayFunction.time_mock_broadcast_to_numpy                                   2.22±0.05μs
[ 75.00%] ··· bench_overrides.ArrayFunction.time_mock_concatenate_duck                                     3.53±0.09μs
[ 83.33%] ··· bench_overrides.ArrayFunction.time_mock_concatenate_many                                        74.1±1μs
[ 91.67%] ··· bench_overrides.ArrayFunction.time_mock_concatenate_mixed                                     4.37±0.2μs
[100.00%] ··· bench_overrides.ArrayFunction.time_mock_concatenate_numpy                                    2.71±0.02μs
```
For the hstack in particular, I see:
```
l = [np.arange(1000), np.arange(1000)]
%timeit np.hstack(l)
# no overrides: 4.04 µs ± 85.2 ns per loop
# master: 10.6 µs ± 108 ns per loop
# this PR: 8.94 µs ± 97.7 ns per loop
# without nested dispatch: 5.95 µs ± 241 ns per loop
```
"With nested dispatch" refers to manually adjusting `hstack()` to call `atleast_1d.__wrapped__` and `concatenate.__wrapped__`, and thus avoid re-invoking `__array_function__` dispatching. This might be a good option for speeding up NumPy functions that invoke other NumPy functions -- but if so, we should also consider exposing `__wrapped__` (or maybe another attribute, such as `implementation`) as part of NumPy's public interface for the benefit of third-party libraries that also want the speed improvement.