@shoyer 2018-12-03
Thanks to some performance tips from @eric-wieser, this latest version is working *much* faster.

Pure Python implementation:
```
$ asv run -b bench_override -E existing -a sample_time=0.5
bench_overrides.ArrayFunction.time_mock_broadcast_to_duck    3.57±0.04μs
bench_overrides.ArrayFunction.time_mock_broadcast_to_numpy   1.96±0.03μs
bench_overrides.ArrayFunction.time_mock_concatenate_duck     3.82±0.08μs
bench_overrides.ArrayFunction.time_mock_concatenate_many         183±3μs
bench_overrides.ArrayFunction.time_mock_concatenate_mixed     7.25±0.2μs
bench_overrides.ArrayFunction.time_mock_concatenate_numpy    2.32±0.01μs
```

C implementation:
```
$ asv run -b bench_override -E existing -a sample_time=0.5
bench_overrides.ArrayFunction.time_mock_broadcast_to_duck    2.64±0.05μs
bench_overrides.ArrayFunction.time_mock_broadcast_to_numpy   1.45±0.03μs
bench_overrides.ArrayFunction.time_mock_concatenate_duck     2.65±0.03μs
bench_overrides.ArrayFunction.time_mock_concatenate_many     2.72±0.03μs
bench_overrides.ArrayFunction.time_mock_concatenate_mixed     4.83±0.2μs
bench_overrides.ArrayFunction.time_mock_concatenate_numpy    1.49±0.01μs
```

These should be compared to my previous results in https://github.com/numpy/numpy/pull/12317#issuecomment-435721980

And here are the results for the `hstack(l)` example:
```
l = [np.arange(1000), np.arange(1000)]
%timeit np.hstack(l)
# no overrides: 4.04 µs ± 85.2 ns per loop
# master: 10.6 µs ± 108 ns per loop
# this PR, Python version: 8.05 µs ± 223 ns per loop
# this PR, C version: 6.74 µs ± 181 ns per loop
# without nested dispatch: 4.69 µs ± 211 ns per loop
```

The bottom line is that overhead for the typical case of checking a single argument that doesn't have an override is now in the range of ~0.7 us, about the level of a single function call. For the typical case of case of no overrides with only a few arguments, the C version is about 30% faster than the Python version. For functions like `concatenate()` with a large number of array arguments (e.g., 1000), the overhead is something like 100x smaller to the point where it doesn't have a noticeable performance impact (previous it was a 4x slow-down).